Improved Searching: Created PopupHelper to reduce Code
