+++ app/src/main/java/com/example/pix/home/fragments/ComposeFragment.java
@@ -32,6 +32,7 @@ import com.example.pix.chat.activities.FriendActivity;
 import com.example.pix.home.activities.HomeActivity;
 import com.example.pix.home.adapters.SearchAdapter;
 import com.example.pix.home.utils.CameraPreview;
+import com.example.pix.home.utils.PopupHelper;
 import com.parse.ParseFile;
 import com.parse.ParseQuery;
 import com.parse.ParseUser;
@@ -95,8 +96,7 @@ public class ComposeFragment extends Fragment {
                     camera.release();
                     if (currCamera == Camera.CameraInfo.CAMERA_FACING_FRONT) {
                         currCamera = Camera.CameraInfo.CAMERA_FACING_BACK;
-                    }
-                    else {
+                    } else {
                         currCamera = Camera.CameraInfo.CAMERA_FACING_FRONT;
                     }
                     camera = Camera.open(currCamera);
@@ -119,52 +119,9 @@ public class ComposeFragment extends Fragment {
         // When we are done taking a picture, go to a new ChatActivity with this new Image ParseFile
         Camera.PictureCallback callback = (bytes, camera) -> {
             image = new ParseFile(bytes);
+            // Create popup to select who to send to
             if (getActivity() instanceof HomeActivity) {
-                LinearLayout container = getActivity().findViewById(R.id.home_container);
-                LayoutInflater layoutInflater = (LayoutInflater) getContext().getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-                View popup = layoutInflater.inflate(R.layout.popup_search, null);
-
-                ImageView close = popup.findViewById(R.id.popup_close);
-                SearchView search = popup.findViewById(R.id.popup_searchView);
-                RecyclerView rvResults = popup.findViewById(R.id.popup_rv);
-
-                // Position popup
-                PopupWindow popupWindow = new PopupWindow(popup, ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.WRAP_CONTENT, true);
-                popupWindow.showAtLocation(container, Gravity.CENTER, 0, 0);
-                search.requestFocus();
-                close.setOnClickListener(view2 -> popupWindow.dismiss());
-
-                List<ParseUser> results = new ArrayList<>();
-                // Tell the adapter whether this adapter will need to handle saved pics(New Snap from ComposeFragment)
-                SearchAdapter adapter = new SearchAdapter(true, getContext(), results);
-                rvResults.setAdapter(adapter);
-                rvResults.setLayoutManager(new LinearLayoutManager(getContext()));
-
-                // When we search, query all users
-                // Once we select a User, the Adapter will handle finding the current User's chat with them,
-                // and if it doesn't exist it will create a new Chat
-                search.setOnQueryTextListener(new SearchView.OnQueryTextListener() {
-                    @Override
-                    public boolean onQueryTextSubmit(String s) {
-                        ParseQuery<ParseUser> q = ParseQuery.getQuery(ParseUser.class);
-                        q.whereStartsWith("username", s);
-                        q.findInBackground((objects, e) -> {
-                            if (e != null) {
-                                Toast.makeText(getContext(), "Error searching!", Toast.LENGTH_SHORT).show();
-                                return;
-                            }
-                            results.clear();
-                            results.addAll(objects);
-                            adapter.notifyDataSetChanged();
-                        });
-                        return false;
-                    }
-
-                    @Override
-                    public boolean onQueryTextChange(String s) {
-                        return false;
-                    }
-                });
+                PopupHelper.createPopup(getActivity(), getContext(), true);
             } else {
                 // Case where this ComposeFragment was called from ChatActivity so we know who to send it to
                 // thus no need for popup
